requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
black==24.8.0
Brotli==1.1.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
crispy-bootstrap4==2023.1
crispy-bootstrap5==0.7
cssselect2==0.7.0
Django==5.0.4
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-environ==0.11.2
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-sms==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fonttools==4.44.0
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.25.0
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
phonenumbers==8.13.36
Pillow==10.0.0
platformdirs==4.3.6
propcache==0.2.0
psycopg==3.2.3
pycparser==2.21
pydyf==0.8.0
pyhtml2pdf==0.0.7
PyJWT==2.8.0
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tinycss2==1.2.1
tkcalendar==1.6.1
tomli==2.0.2
trio==0.23.1
trio-websocket==0.11.1
twilio==9.3.6
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
webdriver-manager==4.0.1
webencodings==0.5.1
whitenoise==6.5.0
wsproto==1.2.0
yarl==1.17.1
zopfli==0.2.3
